+++ CHANGELOG.md
@@ -21,9 +21,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added test for Alexa
 - Added to process local xlsx
 - Added an assets processor to allow downloading media objects from a drive directory
-- Added some unit tests
+- Added some unit tests (Dialogflow)
 
 ## [2.0.1] - 2019-03-05